+++ frontend/jgenesis-gui/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod app;
 mod emuthread;
+mod widgets;